Consolidate defaultPasswordValidator
